feat: add an updateeUser function
