+++ src/index.js
@@ -9,22 +9,6 @@ dotenv.config();
 const PORT = process.env.PORT || 4000;
 let users = [];
 
-// Handle POST reque
-
-// Handle PUT requests
-const handleUpdateUser = (res, userId, body) => {
-    if (!userId || !uuidv4.validate(userId)) {
-        return sendJsonResponse(res, 400, { message: 'Invalid userId' });
-    }
-    const userIndex = users.findIndex((u) => u.id === userId);
-    if (userIndex === -1) {
-        return sendJsonResponse(res, 404, { message: 'User not found' });
-    }
-    const { username, age, hobbies } = JSON.parse(body);
-    users[userIndex] = { ...users[userIndex], username, age, hobbies };
-    sendJsonResponse(res, 200, users[userIndex]);
-};
-
 // Handle DELETE requests
 const handleDeleteUser = (res, userId) => {
     if (!userId || !uuidv4.validate(userId)) {
